committing now to continue working on adding notes and journals into the database

+++ src/main/java/com/example/pickitup/services/dao/NotesDAO.java
@@ -37,7 +37,7 @@ public class NotesDAO {
                 PreparedStatement preparedStatement = connection.prepareStatement(getIDStatement);
                 ResultSet resultSet = preparedStatement.executeQuery()
         ){
-            note.setNoteId(resultSet.getLong("notes_id"));
+            note.setNoteId(resultSet.getInt("notes_id"));
         } catch (SQLException error){
             System.out.println("Error retrieving note id: " + error.getMessage());
         }
@@ -54,8 +54,8 @@ public class NotesDAO {
                 ResultSet resultSet = preparedStatement.executeQuery()
                 ){
             //retrieve the columns and store into variables
-            long note_id = resultSet.getLong("id");
-            long journal_id = resultSet.getLong("journal_id");
+            int note_id = resultSet.getInt("id");
+            int journal_id = resultSet.getInt("journal_id");
             String noteTitle = resultSet.getString("title");
             String noteContent = resultSet.getString("content");
             Note note = new Note(note_id, journal_id, noteTitle, noteContent);
